+++ src/screens/WalletCreatedScreen/WalletCreatedScreen.styles.ts
@@ -3,6 +3,7 @@ import { StyleSheet } from 'react-native';
 export default StyleSheet.create({
 	container: {
 		flex: 1,
+		alignItems: 'center',
 		justifyContent: 'space-between'
 	},
 	image: {
